Test every discovered example
Use example discovery as the CI coverage backstop while preserving tailored registrations. Require test_final() in test mode and document the contract.

Temporarily allow existing deprecation warnings for five newly covered examples. A follow-up will migrate them and remove the allowances.
