+++ newton/tests/test_viewer_loading_splash.py
@@ -3,9 +3,7 @@
 
 import unittest
 from types import SimpleNamespace
-from unittest import mock
 
-import newton.examples
 from newton._src.viewer.viewer_gl import ViewerGL
 
 
@@ -29,12 +27,14 @@ class TestViewerGLLoadingSplashState(unittest.TestCase):
         return viewer
 
     def test_show_sets_active_and_text(self):
+        """Verify showing the splash records its active state and text."""
         viewer = self._make_viewer()
         viewer.show_loading_splash("Loading...")
         self.assertTrue(viewer.gui._loading_splash_active)
         self.assertEqual(viewer.gui._loading_splash_text, "Loading...")
 
     def test_hide_clears_state(self):
+        """Verify hiding the splash clears its active state and text."""
         viewer = self._make_viewer()
         viewer.show_loading_splash("Loading...")
         viewer.hide_loading_splash()
@@ -42,91 +42,12 @@ class TestViewerGLLoadingSplashState(unittest.TestCase):
         self.assertIsNone(viewer.gui._loading_splash_text)
 
     def test_headless_no_gui_is_noop(self):
+        """Verify splash operations are no-ops when no GUI exists."""
         viewer = ViewerGL.__new__(ViewerGL)
         viewer.gui = None
         # Must not raise even though there is no GUI to drive.
         viewer.show_loading_splash("Loading...")
         viewer.hide_loading_splash()
-
-
-class _RecordingViewer:
-    """Stub viewer recording observable calls, used by the lifecycle tests."""
-
-    def __init__(self):
-        self.calls = []
-
-    def show_loading_splash(self, text=None):
-        self.calls.append(("show_loading_splash", text))
-
-    def hide_loading_splash(self):
-        self.calls.append(("hide_loading_splash",))
-
-    def begin_frame(self, t):
-        self.calls.append(("begin_frame", t))
-
-    def end_frame(self):
-        self.calls.append(("end_frame",))
-
-    def is_running(self):
-        return False
-
-    def is_paused(self):
-        return False
-
-    def close(self):
-        self.calls.append(("close",))
-
-
-class TestLoadingSplashLifecycle(unittest.TestCase):
-    """``init()`` shows the splash for visible GL viewers; ``run()`` hides it."""
-
-    def _args(self, **overrides):
-        defaults = {
-            "viewer": "gl",
-            "headless": False,
-            "paused": False,
-            "device": None,
-            "quiet": True,
-            "warp_config": [],
-            "benchmark": False,
-            "realtime": False,
-            "output_path": None,
-            "num_frames": 1,
-            "rerun_address": None,
-            "test": False,
-        }
-        defaults.update(overrides)
-        return SimpleNamespace(**defaults)
-
-    def _run_init(self, args):
-        stub = _RecordingViewer()
-        parser = mock.MagicMock()
-        parser.parse_args.return_value = args
-        with (
-            mock.patch("newton.viewer.ViewerGL", return_value=stub),
-            mock.patch("newton.examples._apply_warp_config"),
-        ):
-            newton.examples.init(parser=parser)
-        return stub
-
-    def test_init_shows_splash_for_visible_gl(self):
-        viewer = self._run_init(self._args())
-        self.assertIn(("show_loading_splash", "Loading..."), viewer.calls)
-
-    def test_init_skips_splash_for_headless(self):
-        viewer = self._run_init(self._args(headless=True))
-        self.assertNotIn(("show_loading_splash", "Loading..."), viewer.calls)
-
-    def test_run_hides_splash(self):
-        viewer = _RecordingViewer()
-        example = SimpleNamespace(
-            viewer=viewer,
-            step=lambda: None,
-            render=lambda: None,
-        )
-        args = SimpleNamespace(test=False)
-        newton.examples.run(example, args)
-        self.assertIn(("hide_loading_splash",), viewer.calls)
 
 
 if __name__ == "__main__":
